Reorganize tests for plugins
