+++ rebasehelper/tests/plugins/test_checkers.py
