+++ rebasehelper/tests/plugins/test_output_tools.py
